requirements.txt
fastapi
uvicorn[standard]
python-multipart
websockets
supabase
python-dotenv
